requirements.txt
torch
transformers==4.41.2
datasets==2.20.0
accelerate==0.31.0
sentencepiece
pyyaml
tqdm
numpy
